added pause sounds and buttons, and fixed some more FMOD errors...
