+++ ProyectoWipStudios/Assets/Plugins/FMOD/Resources/FMODStudioSettings.asset
@@ -39,7 +39,7 @@ MonoBehaviour:
     Value: 0
   LiveUpdateSettings:
   - Platform: 1
-    Value: 1
+    Value: 0
   - Platform: 2
     Value: 0
   - Platform: 20
