+++ ProyectoWipStudios/Assets/Scripts/GameController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c2751e8e8c71438418841777a21cd79d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
