+++ ProyectoWipStudios/Assets/Scripts/SoundManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a40c2122bcb21a549af57ec53031fae7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
