--- ProyectoWipStudios/Assets/TestFmod.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a0bbbbaf0f7bf7648b354344a51b2cd8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
